add cobra cli to the project along with the order grpc client

+++ user-svc/internal/domain/address.go
@@ -1,11 +1,11 @@
 package domain
 
 type Address struct {
-	ID           uint   `json:"id"`
+	ID           int64  `json:"id"`
 	AddressLine1 string `json:"address_line1"`
 	AddressLine2 string `json:"address_line2"`
 	City         string `json:"city"`
 	PostCode     string `json:"post_code"`
 	Country      string `json:"country"`
-	UserId       uint   `json:"user_id"`
+	UserId       int64  `json:"user_id"`
 }
